XRDDEV-2211 generating shared parameters

+++ src/central-server/admin-service/globalconf-generator/src/main/java/org/niis/xroad/cs/admin/globalconf/generator/SharedParameters.java
@@ -120,7 +120,7 @@ class SharedParameters {
     @AllArgsConstructor
     public static class GlobalSettings {
         private List<MemberClass> memberClasses;
-        private Long ocspFreshnessSeconds;
+        private Integer ocspFreshnessSeconds;
     }
 
     @Data
